Fix crash (NPE) caused by changes introduced in "Add reconnect function when network changes" commit
In my previous commit unregisterNetworkCallback was causing a null point exception when called.
This commit fixes the bug, properly unregister the NetworkCallback and set NetworkCallback and
ConnectivityManager null when MusicService is destroyed. I've also added an additional check, now
the code is executed only if ConnectivityManager is not null, in this way even if for some reason
ConnectivityManager is null (in that case we have a problem) the app will not crash if the user
stop the radio streaming service

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>