Settings: hide media notification preference for devices under Android 11
Media notification is a new feature introduced in android 11, here we can set a full media
notification simply setting the session token, and a normal notification when the token is not set.
In older Android version therre's no media notification, so nothing in UI change. But some devices like
Android Auto or wearables needs the token for properly set a working notification.
This commit set the token for all Android versions by default, and hide the option for devices under Android 11
In addition, the get/set color code has been temporary commented, it would be nice to add a new feature and allow the user
to choose the preferred color.

NOTE: DISABLE AT YOUR OWN RISK.

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>