Add a specific AlertDialog for MiUi and EMUI devices
MiUi and EMUI Roms completely changed some android standard behaviours, and one of them is the
way how the OS handles foreground services. In AOSP a foreground service started by the user could run
basically forever (i'm not taking in consideration some cases like low battery or memory, errors, etc),
and this is the correct way to handle this type of service. But Xiaomi and Huawei with their changes
completely broke that function (it's not a bug, the change is intended), and in MiUi and EMUI roms a
foreground service can be killed from system (often from their custom battery service) or from user
simply removing the app from recent apps list (in stock AOSP the app gets closed, but the service is NOT killed).
There's no way to work around this behaviour from the app side, if the app gets killed when in background
the user must manually disable the battery optimisation for the app. Also, for keeping the service alive
when a recent tasks cleaning is performed, the app must be manually pinned in recent app list.

This commit adds an alertdialog (only for MiUi and EMUI devices) with the instructions in case the streaming
service gets killed by system or when app is remover from recent app list.

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>