[HAX] Fix audio sometimes not playing
After some minutes in pause state with the app in background sometimes
play action won't work at all. And there's nothing in logs that helps me actually.
The javascript is not executed and isMediaPlayingZoo remain false. Only ways for
recover the stream are: 1 open the app or 2 press pause/play button again on notification.
This is just an ugly hack, and should be removed as soon as possible

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>