[HACK] Fix stream sometimes not starting when app is in background
When playing a podcast if streaming is paused and app is in background, after a few minutes 105.net site close the connection.
This is normal, the problems starts when player.play() is called: I still don't know why (I haven't inspected scripts on site side yet)
but calling play() does not load again the audio url, so the streaming doesn't start. The strange thing is that pause() do it.
That's how this hack works: when user press play button in notification it simply calls pause() and play() if the streaming is not started after 150 msecs.

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>