[WIP] PiP: support Mediaset Play Videos
As far as i know there is no way for me to check the state of the player, because it's embedded in an iframe and domain policy blocks every js injection attempt, no matter what i try to do.
There's a workaround that may work, but it's an hack and has a significant performance cost, especially on old devices.
For these reasons the onPlay/onPause check has been removed, this fixes PiP mode, but with a drawback: the app will enter in PiP mode even if the video is on pause state, but user cannot interact with it.
The solution would be a MediaSession, but:

1 - Unlike Chrome, WebView lacks MediaSession API, everything must be handled on app side
2 - There's still no way to communicate with the video inside the iframe

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>