TV Fragment: fix progressBar code in onInfo listener
This commit fixes the progress bar sometimes not disappearing even if the video is playing.
Unfortunately MEDIA_INFO_VIDEO_RENDERING_START notify when the player pushes only the VERY first frame, this means that if we go again in buffer state the progressBar will reappear and will never go away.

For this reason i'm forced to drop MediaPlayer.OnInfoListener in favour of MediaPlayer.OnPreparedListener, in this way the progress bar will disappear once the video is ready to play, and won't reappear until fragment is recreated

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>