Settings: move WebViews postVisualStateCallback to experimental options
Thanks to the compat lib postVisualStateCallback is available for all API versions, and if the device does not support it there's a fallback method.
For this reason the default option value now is set to true, and the option in settings is moved under experimental section, as it's no more needed.

NOTE: keep and hide the option instead removing it because there's still a case that could be useful: if postVisualStateCallback is available but for some reason is not working well in user device.

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>