RadioService: fix a memory leak in ConnectivityManager
mConnectivityManager is used in RadioService processStopRequest, where unregisterNetworkCallback is called.
At that point, context may be a dead reference, especially if the app is in background for a few minutes and user press the stop button in notification.
With getApplicationContext we are sure that the reference is recycled only if the application is killed, avoiding the memory leak

TEST: stopped streaming from notification after 10 minutes with the app in background and with the app closed, leakcanary now is happy

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>