RadioService: fix reconnect option for newer Android versions
In latest Android versions we are no more allowed to start a foreground service
when the app is in background. As result, reconnect feature was broken for devices
running android 11 and above, unless screen is on and app is not in background.

TEST: run in Android 14 emulator, start streaming, exit app and turned off screen,
after 22 minutes server close the connection and reconnect feature works as it should

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>